add handle errors for users fucn

+++ db/users_database.py
@@ -1,7 +1,9 @@
+from sqlalchemy import or_
 from sqlalchemy.orm.session import Session
 from db.hash import Hash
 from db.models import DbUser
 from schemas import UserBase
+from fastapi import HTTPException,status
 
 def create_user(db: Session, request: UserBase):
 
@@ -14,6 +16,16 @@ def create_user(db: Session, request: UserBase):
         password = Hash.bcrypt(request.password)
 
     )
+    exist = db.query(DbUser).filter(
+        or_(
+            DbUser.username == request.username,
+            DbUser.email == request.email
+            )
+            ).first()
+    if exist:
+        raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail="Username or email already exists")
+    
+
     db.add(new_user)
     db.commit()
     db.refresh(new_user)
@@ -27,6 +39,8 @@ def get_all_users(db:Session):
 def get_user(db:Session,id:int):
 
     user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="user not found")
     #handle error
 
     return user
@@ -34,8 +48,9 @@ def get_user(db:Session,id:int):
 def delete_user(db:Session, id:int):
 
     user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="user not found")
 
-    #handle error
     db.delete(user)
     db.commit()
